H3M6 Update
Decided to upload my new version of my map early, just in case something terrible happens.

I renamed it to "Hellfire Stronghold" and most of the map was remade for better use of ZDoom features, off with the excesive 3D floors and on with the new portals.
The map is still in beta, and it's still unfinished, it gets especially rough by the end, but it can be completed. Feel free to give criticism if you want, and don't be afraid if I answer with a "I was planning to change that anyways", because there's a chance I was not :P