H3M6 Performance Update
Removed half of the fire spawners and merged all portals from the outside area into a single portal. The frame rate improved considerably on my end.
Also I moved the dagger and scimitar to a secret area, and added a Master Scroll in their original place, and removed the script suspenders placed in interiors.

Also I updated the D'sparil Statue so now it uses the new sprites Captain Toenail made.