H4M8 & Hand Crystal update
Working on the outside area of H4M8, also the Hand Crystal was converted to Zscript and now it can change colors with scripts (with SetActorState). Sorry for the excess of states and else ifs, but I can't think of a more clean way to implement this stuff and my brain hurts.