More dynamic lights & H2M6 update
Dynamic light definitions for the Polymorph Wand, melee weapons and teleport fogs.

Minor update to H2M6: wall torches were lowered, fixed a bug with the torches on the stairs room moving with the floor, raised the sector with the Green Key so you can't get it if you straferun and jump, the gate at the start has been lowered so you have to press the switch to enter, minor visual bugs fixed.